% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cameras.R
\name{parse_cameras}
\alias{parse_cameras}
\title{Parse a list of camera objects into a tibble}
\usage{
parse_cameras(result)
}
\arguments{
\item{result}{List returned by \code{nims_request("/cameras")}.}
}
\value{
A tibble.
}
\description{
Parse a list of camera objects into a tibble
}
\keyword{internal}
